tests/wpt/web-platform-tests/performance-timeline/tentative/include-frames-originA-A.html

<!DOCTYPE html>

<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>

<body>
<script>
  const loadSameOriginChildFrame = () => {
    return new Promise(resolve => {

      const childFrame = document.createElement('iframe');

      childFrame.addEventListener("load", () => {
        resolve();
      });

      childFrame.src = "../resources/child-frame.html";

      document.body.appendChild(childFrame);
    });
  }

  promise_test(async () => {
    performance.clearResourceTimings();

    // Load a child frame.
    await loadSameOriginChildFrame();

    const entries = performance.getEntries(true);

    const navigationEntries = performance.getEntriesByType('navigation', true);

    const markedEntries = performance.getEntriesByName('entry-name', undefined, true);

    // 3 entries for parent, 2 for child.
    assert_equals(entries.length, 5, 'Total entries should be 5.');

    // 1 entry for parent, 1 for child.
    assert_equals(navigationEntries.length, 2, 'Navigation entries should be 2.');

    // 1 entry for child.
    assert_equals(markedEntries.length, 1, 'Mark entries should be 1.');
  }, 'GetEntries of a document of origin A and its child frame of origin A.');
</script>
</body>
